bug fixed a print

+++ examples/cleanRL_cartpole_ppo.py
@@ -41,7 +41,7 @@ def training_loop(env):
             model.train_net()
 
         if n_epi%print_interval==0 and n_epi!=0:
-            print("# of episode :{}, avg score : {:.1f}".format(n_epi, score/print_interval))
+            print("# of episode :{}, avg score : {:.1f}".format(n_epi, float(score/print_interval)))
             score = 0.0
